Chain Lightning

+++ Assets/Interfaces/IHasHealth.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IHasHealth
 {
-    public void OnHit(int damage, GameObject attacker);
+    public void OnHit(float damage, GameObject attacker);
     public void Die();
 }
