+++ Assets/Scripts/TestPlayer.cs
@@ -2,14 +2,14 @@ using UnityEngine;
 
 public class TestPlayer : MonoBehaviour, IHasHealth
 {
-    public int health = 6;
+    public float health = 6;
 
     public void Die()
     {
         Destroy(gameObject);
     }
 
-    public void OnHit(int damage, GameObject attacker)
+    public void OnHit(float damage, GameObject attacker)
     {
         EventManager.Instance.OnHit(gameObject, attacker); 
         health -= damage;
